editor/libeditor/tests/test_bug645914.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=645914
-->
<head>
  <title>Test for Bug 645914</title>
  <script src="/tests/SimpleTest/SimpleTest.js"></script>
  <script src="/tests/SimpleTest/EventUtils.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=645914">Mozilla Bug 645914</a>
<p id="display"></p>
<div id="content">
<textarea>foo
bar</textarea>
</div>
<pre id="test">
<script type="application/javascript">

/** Test for Bug 645914 **/
SimpleTest.waitForExplicitFinish();
SimpleTest.waitForFocus(function() {
  SpecialPowers.pushPrefEnv({"set": [["layout.word_select.eat_space_to_next_word", true],
                                     ["browser.triple_click_selects_paragraph", false]]}, startTest);
});
function startTest() {
  var textarea = document.querySelector("textarea");
  textarea.selectionStart = textarea.selectionEnd = 0;

  // Simulate a double click on foo
  synthesizeMouse(textarea, 5, 5, {clickCount: 2});

  ok(true, "Testing word selection");
  is(textarea.selectionStart, 0, "The start of the selection should be at the beginning of the text");
  is(textarea.selectionEnd, 3, "The end of the selection should not include a newline character");

  textarea.selectionStart = textarea.selectionEnd = 0;

  // Simulate a triple click on foo
  synthesizeMouse(textarea, 5, 5, {clickCount: 3});

  ok(true, "Testing line selection");
  is(textarea.selectionStart, 0, "The start of the selection should be at the beginning of the text");
  is(textarea.selectionEnd, 3, "The end of the selection should not include a newline character");

  textarea.selectionStart = textarea.selectionEnd = 0;
  textarea.value = "Very very long value which would eventually overflow the visible section of the textarea";

  // Simulate a quadruple click on Very
  synthesizeMouse(textarea, 5, 5, {clickCount: 4});

  ok(true, "Testing paragraph selection");
  is(textarea.selectionStart, 0, "The start of the selection should be at the beginning of the text");
  is(textarea.selectionEnd, textarea.value.length, "The end of the selection should be the end of the paragraph");

  SimpleTest.finish();
}
</script>
</pre>
</body>
</html>
